clean: Move all styles to separate folder
