+++ src/components/CTA.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "../style";
+import styles from "../styles/style";
 
 const CTA = () => {
   return (
