+++ src/components/Footer.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "../style";
+import styles from "../styles/style";
 import { logo } from "../assets";
 import { footerLinks, socialMedia } from "../constants";
