+++ src/components/GetStarted.jsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "../style";
+import styles from "../styles/style";
 import { arrowUp } from "../assets";
 
 const GetStarted = () => {
