+++ src/main.jsx
@@ -1,7 +1,9 @@
 import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App'
-import './index.css'
+import './styles/index.css'
+import './styles/tables.css'
+import './styles/discounts.css'
 
 ReactDOM.createRoot(document.getElementById('root')).render(
   <React.StrictMode>
